include gch defaults, move gch gaussian_model_base
