+++ floris/simulation/wake_velocity/gaussianModels/gauss_legacy.py
@@ -13,18 +13,23 @@
 import numpy as np
 from ....utilities import cosd, sind, tand, setup_logger
 from ..base_velocity_deficit import VelocityDeficit
-from .gaussian_model_ish import GaussianModel
+from .gaussian_model_base import GaussianModel
 
 
-class LegacyGauss(VelocityDeficit):
+class LegacyGauss(GaussianModel):
     default_parameters = {
         'ka': 0.38,
         'kb': 0.004,
         'alpha': 0.58,
-        'beta': 0.077
+        'beta': 0.077,
+        'calculate_VW_velocities':False,
+        'use_yaw_added_recovery':False,
+        'yaw_recovery_alpha':0.03,
+        'eps_gain':0.3
     }
 
     def __init__(self, parameter_dictionary):
+
         super().__init__(parameter_dictionary)
         self.logger = setup_logger(name=__name__)
 
@@ -38,6 +43,13 @@ class LegacyGauss(VelocityDeficit):
         # wake expansion parameters
         self.ka = model_dictionary["ka"]
         self.kb = model_dictionary["kb"]
+
+        # GCH Parameters
+        self.calculate_VW_velocities = model_dictionary["calculate_VW_velocities"]
+        self.use_yaw_added_recovery = model_dictionary["use_yaw_added_recovery"]
+        self.yaw_recovery_alpha = model_dictionary["yaw_recovery_alpha"]
+        self.eps_gain = model_dictionary["eps_gain"]
+
 
     def function(self, x_locations, y_locations, z_locations, turbine, turbine_coord, deflection_field, flow_field):
         # veer (degrees)
